backend/app/pyproject.toml
[tool.poetry]
name = "fast-api-ec-demo-server"
version = "0.1.0"
description = ""
authors = ["YuanShiDong <yuan@webimpact.co.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.13.3"
fastapi = "^0.63.0"
python-multipart = "^0.0.5"
requests = "^2.25.1"
SQLAlchemy = "^1.3.22"
inflect = "^5.0.2"
python-dotenv = "^0.15.0"
phonenumbers = "^8.12.15"
email-validator = "^1.1.2"
alembic = "^1.4.3"
PyMySQL = "^1.0.1"
PyYAML = "^5.3.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
